requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
SQLAlchemy~=1.4.45
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==0.21.1
pydantic~=2.8.2
pydantic-settings==2.3.4
passlib~=1.7.4
starlette~=0.37.2
pyjwt~=2.8.0
passlib[bcrypt]~=1.7.4
httpx~=0.27.0